requirements.txt
folium==0.19.5
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
shapely==2.0.7
requests==2.32.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dotenv==1.1.0
PyGithub==2.6.1
